Move indexByte to an internal package
This allows both the ctcp package and the irc package to share the code, which
should allow the build to succeed for 1.1 again.
