+++ .travis.yml
@@ -14,4 +14,3 @@ script:
 matrix:
   allow_failures:
     - go: 1.0
-    - go: 1.1
