+++ message.go
@@ -7,6 +7,8 @@ package irc
 import (
 	"bytes"
 	"strings"
+
+	"gopkg.in/sorcix/irc.v2/internal"
 )
 
 // Various constants used for formatting IRC messages.
@@ -52,8 +54,8 @@ func ParsePrefix(raw string) (p *Prefix) {
 
 	p = new(Prefix)
 
-	user := indexByte(raw, prefixUser)
-	host := indexByte(raw, prefixHost)
+	user := internal.IndexByte(raw, prefixUser)
+	host := internal.IndexByte(raw, prefixHost)
 
 	switch {
 
@@ -118,7 +120,7 @@ func (p *Prefix) IsHostmask() bool {
 
 // IsServer returns true if this prefix looks like a server name.
 func (p *Prefix) IsServer() bool {
-	return len(p.User) <= 0 && len(p.Host) <= 0 // && indexByte(p.Name, '.') > 0
+	return len(p.User) <= 0 && len(p.Host) <= 0 // && internal.IndexByte(p.Name, '.') > 0
 }
 
 // writeTo is an utility function to write the prefix to the bytes.Buffer in Message.String().
@@ -179,7 +181,7 @@ func ParseMessage(raw string) (m *Message) {
 	if raw[0] == prefix {
 
 		// Prefix ends with a space.
-		i = indexByte(raw, space)
+		i = internal.IndexByte(raw, space)
 
 		// Prefix string must not be empty if the indicator is present.
 		if i < 2 {
@@ -193,7 +195,7 @@ func ParseMessage(raw string) (m *Message) {
 	}
 
 	// Find end of command
-	j = i + indexByte(raw[i:], space)
+	j = i + internal.IndexByte(raw[i:], space)
 
 	// Extract command
 	if j > i {
@@ -209,7 +211,7 @@ func ParseMessage(raw string) (m *Message) {
 	j++
 
 	// Find prefix for trailer
-	i = indexByte(raw[j:], prefix)
+	i = internal.IndexByte(raw[j:], prefix)
 
 	if i < 0 || raw[j+i-1] != space {
